added fake-auth middleware & updated docs
