+++ README.md
@@ -1,3 +1,35 @@
 # express-sessions-redis
 
 Managing Node.js - Express Sessions with Redis
+
+This is a NodeJs demo project, which explain the working of Redis caching with Express session.
+
+If you have any questions, you can mail me at `mail@ankitpant.com` or `ankit12.pant@gmail.com`.
+
+## Development server
+
+Run `yarn start` for a dev server. Navigate to `http://localhost:8080/`. The app will automatically reload if you change any of the source files.
+
+## Build
+
+Run `yarn build` to build the project. The build artifacts will be stored in the `dist/` directory.
+
+## Running unit tests
+
+@Todo
+Run `yarn test` to execute the unit tests via [Jest](https://jestjs.io/)
+or `yarn test:watch` to watch the changes in the unit testing while you code them.
+This project will also generate the code coverage which can be seen by hosting/deploying
+coverage folder.
+
+## Built With ❤️ and
+
+- [NodeJs](https://nodejs.org/en/)
+- [Express](expressjs.com)
+- [Redis](https://redis.io/)
+- [TypeScript](http://www.typescriptlang.org/)
+- [Pug](https://pugjs.org/api/getting-started.html)
+
+## Author
+
+[Ankit Pant](https://ankitpant.com)
