+++ src/middleware/fake-auth.middleware.ts
@@ -1,0 +1,17 @@
+export async function useFakeAuthAuthMiddleware(req: any, res: any, next: any) {
+  try {
+    // If this user already has a session, return
+    if (!req.session?.user) {
+      // Use fake user
+      req.session.user = {
+        username: 'dummyName',
+        displayName: 'Dummy User',
+        email: 'dummy.user@domain.com',
+      };
+    }
+  } catch (error) {
+    console.error(error);
+  } finally {
+    return next();
+  }
+}
